Update 2d car setup to work
Expose all the tuning so I can adjust it in the asset.

Some of these values are orders of magnitude off so they weren't working
properly. Same with the "close enough to zero" speed in apply_friction.

You can feel the driftyness from the low traction and it plays pretty
well.